+++ scripts/normalize_names.py
@@ -3,6 +3,7 @@
 Uses names from the Glyphs database and replaces names like "uni04B5"
 with human-readable names like "tetsecyrillic"
 """
+
 from __future__ import annotations
 
 import sys
@@ -12,31 +13,48 @@ from pathlib import Path
 import requests
 from glyphsLib import GSComponent, GSFont, GSGlyph
 
-GIST_ID = "635b753408f9b3abf949eb82b1e40b28"
-GLYPHS_DATA_FILE = "glyphs-data.xml"
-GLYPHS_DATA_URL = f"https://gist.githubusercontent.com/mishamyrt/{GIST_ID}/raw/{GLYPHS_DATA_FILE}"
-GLYPHS_DATA_DIR = "/tmp/glyphs-data"
-GLYPHS_DATA_PATH = f"{GLYPHS_DATA_DIR}/{GLYPHS_DATA_FILE}"
+GITHUB_REPO = "schriftgestalt/GlyphsInfo"
+GITHUB_REPO_BRANCH = "Glyphs3"
+GLYPHS_DATA_FILE_NAME = "GlyphData.xml"
+GLYPHS_DATA_URL = (
+    f"https://raw.githubusercontent.com/{GITHUB_REPO}"
+    f"/refs/heads/{GITHUB_REPO_BRANCH}/{GLYPHS_DATA_FILE_NAME}"
+)
+GLYPHS_DATA_PATH = Path("/tmp") / GLYPHS_DATA_FILE_NAME
 
 Rename = tuple[str, str]
 
-def _get_glyph_names():
-    """Get names from the Glyphs database"""
-    Path(GLYPHS_DATA_DIR).mkdir(parents=True, exist_ok=True)
-    if GLYPHS_DATA_FILE not in Path(GLYPHS_DATA_DIR).iterdir():
-        print(f"Downloading data from {GLYPHS_DATA_URL}...")
-        response = requests.get(GLYPHS_DATA_URL, timeout=10)
-        with Path(GLYPHS_DATA_PATH).open("wb") as file:
-            file.write(response.content)
-    else:
-        print("Using cached data")
-    tree = ET.parse(GLYPHS_DATA_PATH)
-    root = tree.getroot()
+
+def _get_glyph_data() -> bytes:
+    """Get glyph data from the Glyphs database.
+    Downloads and caches the data if it's not already present.
+    Returns bytes of the XML file.
+    """
+    if GLYPHS_DATA_PATH.exists():
+        return GLYPHS_DATA_PATH.read_bytes()
+
+    print(f"Downloading {GLYPHS_DATA_FILE_NAME}...")
+    response = requests.get(GLYPHS_DATA_URL, timeout=10)
+    if response.status_code != 200:
+        print(f"Failed to download {GLYPHS_DATA_FILE_NAME} from {GLYPHS_DATA_URL}")
+        print("Status code:", response.status_code)
+        print("Text:", response.text)
+        exit(1)
+
+    content = response.content
+    GLYPHS_DATA_PATH.write_bytes(content)
+    return content
+
+
+def _get_glyph_names(glyph_data: bytes) -> dict[str, str]:
+    """Get a dictionary of unicode values to names from the Glyphs database."""
+    tree = ET.fromstring(glyph_data)
     names = {}
-    for elem in root.iter():
+    for elem in tree.iter():
         if elem.tag == "glyph" and "unicode" in elem.attrib:
             names[elem.attrib["unicode"]] = elem.attrib["name"]
     return names
+
 
 def _deep_rename(font: GSFont, renames: list[Rename]) -> None:
     """Renames the glyph"""
@@ -49,9 +67,9 @@ def _deep_rename(font: GSFont, renames: list[Rename]) -> None:
                     if isinstance(shape, GSComponent) and shape.name == original:
                         shape.name = target
 
+
 def _find_renames(
-    font: GSFont,
-    known_glyphs: dict[str, str]
+    font: GSFont, known_glyphs: dict[str, str]
 ) -> tuple[list[Rename], list[GSGlyph]]:
     """Renames uni-prefixed glyphs"""
     renames = []
@@ -79,9 +97,13 @@ def _find_renames(
     renames += additional_renames
     return sorted(renames, key=lambda rename: rename[0]), unknown
 
+
 def normalize_names(input_path: str, output_path: str):
     """Normalize the names of glyphs in the input file"""
-    known_glyphs = _get_glyph_names()
+    print("Getting glyph data...")
+    glyph_data = _get_glyph_data()
+    known_glyphs = _get_glyph_names(glyph_data)
+
     print("Finding wrong names...")
     font = GSFont(input_path)
     renames, unknown = _find_renames(font, known_glyphs)
@@ -104,8 +126,16 @@ def normalize_names(input_path: str, output_path: str):
     font.save(output_path)
     print(f"Saved to {output_path}")
 
-if __name__ == "__main__":
-    if len(sys.argv) < 3:
-        print(f"Usage: {sys.argv[0]} <input_file> <output_file>")
+
+def main():
+    """Main entrypoint"""
+    if len(sys.argv) < 2:
+        print(f"Usage: {sys.argv[0]} <input_file> [<output_file>]")
         sys.exit(1)
-    normalize_names(sys.argv[1], sys.argv[2])
+    input_path = sys.argv[1]
+    output_path = sys.argv[2] if len(sys.argv) > 2 else input_path
+    normalize_names(input_path, output_path)
+
+
+if __name__ == "__main__":
+    main()